:bug: fix #2 with improved readability
This is the improved version of an earlier attempt of mine.
Now there should be no problem anymore with characters or code that
breaks the script.
The only unwanted case i can think of at the moment is injection of fake
commits.
Even if someone might try to break it, the message should just be cut
down.